Complete rewrite of nestedjoin
