+++ pilot/pkg/serviceregistry/kube/controller/ambient/ambientindex.go
@@ -18,7 +18,6 @@ import (
 	"net/netip"
 	"strings"
 
-	"github.com/davecgh/go-spew/spew"
 	corev1 "k8s.io/api/core/v1"
 	discovery "k8s.io/api/discovery/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -615,7 +614,6 @@ func (a *index) Lookup(key string) []model.AddressInfo {
 		for _, w := range a.workloads.ByServiceKey.Lookup(svc.ResourceName()) {
 			res = append(res, w.AsAddress)
 		}
-		log.Infof("Service lookup returned the following WDS addresses: %s", spew.Sprint(res))
 		return res
 	}
