+++ pkg/kube/krt/nestedjoin.go
@@ -33,14 +33,13 @@ type collectionLister[T any] interface {
 // to any keys in the collection itself (otherwise, different handlers will
 // read/mutate the same state and see a different view of the world).
 type nestedjoin[T any] struct {
+	id                       collectionUID
+	collectionName           string
 	collections              internalCollection[Collection[T]]
 	synced                   <-chan struct{}
 	syncer                   Syncer
 	collectionChangeHandlers *collectionChangeHandlers[T]
 	metadata                 Metadata
-	// internal indexes
-	indexes map[string]joinCollectionIndex[T]
-	*mergejoin[T]
 
 	// Keep a stop channel so we can check if nested collections are synced
 	stop <-chan struct{}
@@ -51,32 +50,54 @@ type handlerLocalKey[T any] struct {
 	handlerID uint64
 }
 
-type joinCollectionIndex[T any] struct {
-	extract func(o T) []string
-	index   map[string]sets.Set[handlerLocalKey[T]]
-	parent  *mergejoin[T]
-}
-
 type collectionChangeHandlers[T any] struct {
 	handlers []func(collectionChangeEvent[T])
 	sync.RWMutex
-}
-
-func (j *nestedjoin[T]) getCollections() []Collection[T] {
-	// This is used by the collection lister to get the collections for this join
-	// so it can be used in a nested join.
-	return j.collections.List()
 }
 
 func (j *nestedjoin[T]) Metadata() Metadata {
 	return j.metadata
 }
 
+func (j *nestedjoin[T]) List() []T {
+	var res []T
+	var found sets.String
+	first := true
+	for _, c := range j.collections.List() {
+		objs := c.List()
+		// As an optimization, take the first (non-empty) result as-is without copying
+		if len(objs) > 0 && first {
+			res = objs
+			first = false
+			found = sets.NewWithLength[string](len(objs))
+			for _, i := range objs {
+				found.Insert(GetKey(i))
+			}
+		} else {
+			// After the first, safely merge into the result
+			for _, i := range objs {
+				key := GetKey(i)
+				if !found.InsertContains(key) {
+					// Only keep it if it is the first time we saw it, as our merging mechanism is to keep the first one
+					res = append(res, i)
+				}
+			}
+		}
+	}
+	return res
+}
+
+func (j *nestedjoin[T]) GetKey(k string) *T {
+	for _, c := range j.collections.List() {
+		if r := c.GetKey(k); r != nil {
+			return r
+		}
+	}
+	return nil
+}
+
 // nolint: unused // (not true, its to implement an interface)
 func (j *nestedjoin[T]) index(name string, extract func(o T) []string) indexer[T] {
-	if j.merge != nil {
-		return j.mergejoin.index(name, extract)
-	}
 	ji := &dynamicJoinIndexer[T]{indexers: make(map[collectionUID]indexer[T])}
 	for _, c := range j.collections.List() {
 		ic := c.(internalCollection[T])
@@ -118,17 +139,15 @@ func (j *nestedjoin[T]) handleCollectionChangeEvent(
 	djhr *dynamicJoinHandlerRegistration,
 	e collectionChangeEvent[T],
 	handler func(o []Event[T]),
-	handle *registrationHandle,
 ) {
 	djhr.Lock()
 	defer djhr.Unlock()
-	handlerID := handle.id
 
 	switch e.eventType {
 	case collectionMembershipEventAdd:
-		log.Infof("NestedJoinCollection: Adding collection %s with uid %d in handler %d", e.collectionValue.name(), e.collectionValue.uid(), handlerID)
+		log.Infof("NestedJoinCollection: Adding collection %s with uid %d", e.collectionValue.name(), e.collectionValue.uid())
 		// We always want to send events for existing state when a new collection is added
-		reg := e.collectionValue.RegisterBatch(j.handleInnerCollectionEvent(func(o []Event[T]) {
+		reg := e.collectionValue.RegisterBatch(func(o []Event[T]) {
 			djhr.RLock()
 			defer djhr.RUnlock()
 			if _, ok := djhr.removes[e.collectionValue.uid()]; !ok {
@@ -136,7 +155,7 @@ func (j *nestedjoin[T]) handleCollectionChangeEvent(
 				return
 			}
 			handler(o)
-		}, handle), true) // Always run existing state
+		}, true) // Always run existing state
 		djhr.removes[e.collectionValue.uid()] = reg.UnregisterHandler
 		djhr.syncers[e.collectionValue.uid()] = reg
 	case collectionMembershipEventDelete:
@@ -160,166 +179,24 @@ func (j *nestedjoin[T]) handleCollectionChangeEvent(
 		delete(djhr.syncers, e.collectionValue.uid())
 
 		// Now we must send a final set of remove events for each object in the collection
-		var events []Event[T]
 		oldItems := e.collectionValue.List()
-		// Short circuit; send delete events for all items in the collection
-		if j.merge == nil {
-			events := slices.Map(oldItems, func(i T) Event[T] {
-				return Event[T]{Old: &i, Event: controllers.EventDelete}
-			})
-			handler(events)
-			return
-		}
-
-		// We're merging so this is a bit more complicated
-		items := sets.NewWithLength[Key[T]](len(oldItems))
-		// First loop through the collection to get the deleted items by their keys
-		for _, c := range oldItems {
-			key := getTypedKey(c)
-			items.Insert(key)
-		}
-
-		// Now loop through the keys and compare them to our current list of collections
-		// to see if it's actually deleted
-		for key := range maps.SeqStable(items) {
-			keyString := string(key)
-			res := j.calculateMerged(keyString)
-			// Always update the cache on a collection delete
-			entry := j.updateMergeCache(keyString, handlerID, res)
-			var e Event[T]
-			// If the result is nil, then it was deleted
-			if res == nil {
-				// Send a delete event for the merged version of this key
-				// Use the merge of the old items as the old value
-				e = Event[T]{Old: entry.prev, Event: controllers.EventDelete}
-			} else {
-				// There are some versions of this key still in the overall collection
-				// send an update with the new merged version and the old version from
-				// the cache
-				e = Event[T]{Old: entry.prev, New: res, Event: controllers.EventUpdate}
-			}
-
-			j.updateIndex(e, handlerLocalKey[T]{
-				Key:       key,
-				handlerID: handlerID,
-			})
-			events = append(events, e)
-			log.Infof("Merged event %s due to collection delete for key %s in collection %s", e.Event, keyString, j.collectionName)
-		}
+		events := slices.Map(oldItems, func(i T) Event[T] {
+			return Event[T]{Old: &i, Event: controllers.EventDelete}
+		})
 		handler(events)
-	case collectionMembershipEventUpdate:
-		// Get all of the elements in the old collection
-		oldItems := e.oldCollectionValue.List()
-		// Convert it to a map for easy lookup
-		oldItemsMap := make(map[Key[T]]T, len(oldItems))
-		if j.merge == nil {
-			// Short-circuit; send update events for all items in the collection
-			events := slices.Map(oldItems, func(i T) Event[T] {
-				return Event[T]{Old: &i, Event: controllers.EventUpdate, New: j.GetKey(GetKey(i))}
-			})
-			handler(events)
-			return
-		}
-		for _, i := range oldItems {
-			key := getTypedKey(i)
-			oldItemsMap[key] = i
-		}
-		// Now loop through the new collection and compare it to the old one
-		seen := sets.NewWithLength[string](len(oldItems))
-		finalEvents := make([]Event[T], 0, len(oldItems))
-		for _, i := range e.collectionValue.List() {
-			key := getTypedKey(i)
-			// If we see it in the old collection, then it's an update
-			if oldItem, ok := oldItemsMap[key]; ok {
-				seen.Insert(string(key))
-				// Don't need to pass i since the collection still exists and has been updated
-				merged := j.calculateMerged(string(key))
-				if merged == nil {
-					// This shouldn't happen, log it
-					log.Warnf("NestedJoinCollection: Merged item %v is nil after a collection update. Falling back to collection specific version", key)
-					merged = &i
-				} else {
-					// Update the cache with the new merged version
-					entry := j.updateMergeCache(string(key), handlerID, merged)
-					oldItem = ptr.OrEmpty(entry.prev)
-				}
-				// Send an update event for the merged version of this key
-				finalEvents = append(finalEvents, Event[T]{Old: &oldItem, New: merged, Event: controllers.EventUpdate})
-				// Delete it from the old items map
-				delete(oldItemsMap, key)
-			} else {
-				if seen.Contains(string(key)) {
-					// This is a duplicate item in the new collection, skip it
-					log.Warnf("NestedJoinCollection: Duplicate item %v in updated collection, skipping", key)
-					continue
-				}
-				// This is a new item
-				finalEvents = append(finalEvents, Event[T]{New: &i, Event: controllers.EventAdd})
-			}
-		}
-		// Now loop through the old items map and send delete events for any items that
-		// are no longer in the new collection
-		for _, i := range maps.SeqStable(oldItemsMap) {
-			finalEvents = append(finalEvents, Event[T]{Old: &i, Event: controllers.EventDelete})
-		}
+		return
 
-		// Update the indexes
-		for _, e := range finalEvents {
-			j.updateIndex(e, handlerLocalKey[T]{
-				Key:       getTypedKey(e.Latest()),
-				handlerID: handlerID,
-			})
-		}
-		handler(finalEvents)
+	case collectionMembershipEventUpdate:
+		oldItems := e.oldCollectionValue.List()
+		events := slices.Map(oldItems, func(i T) Event[T] {
+			return Event[T]{Old: &i, Event: controllers.EventUpdate, New: j.GetKey(GetKey(i))}
+		})
+		handler(events)
+		return
 	}
 }
 
 func (j *nestedjoin[T]) RegisterBatch(f func(o []Event[T]), runExistingState bool) HandlerRegistration {
-	// Create a unique handler ID for this context
-	h := &registrationHandle{
-		id: globalUIDCounter.Inc(),
-	}
-	return j.registerBatchForHandler(f, runExistingState, h)
-}
-
-func (j *nestedjoin[T]) registerBatchForHandler(f func(o []Event[T]), runExistingState bool, handle *registrationHandle) HandlerRegistration {
-	if j.merge == nil {
-		return j.registerBatchUnmerged(f, runExistingState)
-	}
-	j.Lock()
-	defer j.Unlock()
-	// Save the handler ID
-	j.handlers.Insert(handle.id)
-	syncers := make(map[collectionUID]Syncer)
-	removes := map[collectionUID]func(){}
-	// Only init this new handler with the collection's state if we're not running existing state.
-	// Handlers that run existing state will catch up once they get all of their events.
-	if !runExistingState && handle.id != uint64(j.uid()) {
-		j.maybeInitMergeCacheForHandlerLocked(handle.id)
-	}
-
-	for _, c := range j.collections.List() {
-		ic := c.(internalCollection[T])
-		reg := c.RegisterBatch(j.handleInnerCollectionEvent(f, handle), runExistingState)
-		removes[ic.uid()] = reg.UnregisterHandler
-		syncers[ic.uid()] = reg
-	}
-	djhr := &dynamicJoinHandlerRegistration{
-		syncers:   syncers,
-		removes:   removes,
-		handlerID: handle.id,
-	}
-
-	// We register to get notified if a collection within our set of collections is modified
-	j.registerCollectionChangeHandler(func(e collectionChangeEvent[T]) {
-		j.handleCollectionChangeEvent(djhr, e, f, handle)
-	})
-
-	log.Infof("NestedJoinCollection: Registered handler %d for collection %s", handle.id, j.collectionName)
-	return djhr
-}
-
-func (j *nestedjoin[T]) registerBatchUnmerged(f func(o []Event[T]), runExistingState bool) HandlerRegistration {
 	syncers := make(map[collectionUID]Syncer)
 	removes := map[collectionUID]func(){}
 	for _, c := range j.collections.List() {
@@ -448,17 +325,10 @@ func (j *nestedjoin[T]) registerCollectionChangeHandler(h func(e collectionChang
 }
 
 // NestedJoinCollection creates a new collection of collections of T. Duplicate keys across the collections will *not*
-// be merged (see NestedJoinWithMergeCollection for that) and access operations (e.g. GetKey and List) will perform a
+// be merged (see NestedJoinWithMergeCollection2 for that) and access operations (e.g. GetKey and List) will perform a
 // best effort stable ordering of the list of elements returned; however, this ordering will not be persistent across
 // istiod restarts.
 func NestedJoinCollection[T any](collections Collection[Collection[T]], opts ...CollectionOption) Collection[T] {
-	return NestedJoinWithMergeCollection(collections, nil, opts...)
-}
-
-// NestedJoinWithMergeCollection creates a new collection of collections of T that combines duplicates with the specified merge function.
-// The merge function *cannot* assume a stable ordering of the list of elements passed to it. Therefore, access operations (e.g. GetKey and List) will
-// will only be deterministic if the merge function is deterministic. The merge function should return nil if no value should be returned.
-func NestedJoinWithMergeCollection[T any](collections Collection[Collection[T]], merge func(ts []T) *T, opts ...CollectionOption) Collection[T] {
 	o := buildCollectionOptions(opts...)
 	if o.name == "" {
 		o.name = fmt.Sprintf("NestedJoin[%v]", ptr.TypeName[T]())
@@ -468,7 +338,10 @@ func NestedJoinWithMergeCollection[T any](collections Collection[Collection[T]],
 	synced := make(chan struct{})
 
 	j := &nestedjoin[T]{
-		synced:      synced,
+		synced:         synced,
+		id:             nextUID(),
+		collectionName: o.name,
+
 		collections: ics,
 		syncer: channelSyncer{
 			name:   o.name,
@@ -477,18 +350,8 @@ func NestedJoinWithMergeCollection[T any](collections Collection[Collection[T]],
 		collectionChangeHandlers: &collectionChangeHandlers[T]{
 			handlers: make([]func(collectionChangeEvent[T]), 0),
 		},
-		mergejoin: &mergejoin[T]{
-			id:             nextUID(),
-			mergeCache:     make(map[mergeCacheKey]mergeCacheEntry[T]),
-			indexes:        make(map[string]joinCollectionIndex[T]),
-			merge:          merge,
-			collectionName: o.name,
-			handlers:       sets.New[uint64](),
-		},
 		stop: o.stop,
 	}
-
-	j.mergejoin.collections = j
 
 	if o.metadata != nil {
 		j.metadata = o.metadata
@@ -508,40 +371,22 @@ func NestedJoinWithMergeCollection[T any](collections Collection[Collection[T]],
 			for _, h := range j.collectionChangeHandlers.handlers {
 				switch e.Event {
 				case controllers.EventAdd:
-					if e.New == nil {
-						log.Warnf("Event %v has no new value", e.Event)
-						continue
-					}
 					h(collectionChangeEvent[T]{eventType: collectionMembershipEventAdd, collectionValue: any(*e.New).(internalCollection[T])})
-				// TODO: What does a collection/informer update even look like?
-				// I'm assuming it's something like a change in informer filter
-				// or something and the uid stays the same.
 				case controllers.EventUpdate:
-					if e.New == nil || e.Old == nil {
-						log.Warnf("Event %v either has no New value or no Old value", e.Event)
-						continue
-					}
+					// TODO: What does a collection/informer update even look like?
+					// I'm assuming it's something like a change in informer filter
+					// or something and the uid stays the same.
 					h(collectionChangeEvent[T]{
 						eventType:          collectionMembershipEventDelete,
 						collectionValue:    any(*e.New).(internalCollection[T]),
 						oldCollectionValue: any(*e.Old).(internalCollection[T]),
 					})
 				case controllers.EventDelete:
-					if e.Old == nil {
-						log.Warnf("Event %v has no old value", e.Event)
-						continue
-					}
 					h(collectionChangeEvent[T]{eventType: collectionMembershipEventDelete, collectionValue: any(*e.Old).(internalCollection[T])})
 				}
 			}
 		}
 	}, false)
-
-	j.registerBatchForHandler(func(_ []Event[T]) {
-
-	}, true, &registrationHandle{
-		id: uint64(j.id),
-	})
 
 	go func() {
 		// Need to make sure the outer collection is synced first
@@ -556,8 +401,5 @@ func NestedJoinWithMergeCollection[T any](collections Collection[Collection[T]],
 		close(synced)
 		log.Infof("%v synced", o.name)
 	}()
-
-	log.Infof("NestedJoinCollection: Created collection %s with uid %d", j.collectionName, j.uid())
-
 	return j
 }
